% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbackward.R
\name{fastbackward}
\alias{fastbackward}
\title{Performs Fast Backward Elimination by AIC}
\usage{
fastbackward(
  object,
  scope,
  scale = 0,
  trace = 1,
  keep = NULL,
  steps = 1000,
  k = 2,
  ...
)
}
\arguments{
\item{object}{an object representing a model of an appropriate class
(mainly "\code{lm}" and "\code{glm}"). This is used as the initial model in the stepwise search.}

\item{scope}{defines the range of models examined in the stepwise search. This
should be missing or a single formula. If a formula is included, all of the
components on the right-hand-side of the formula are always included in the model.
If missing, then only the intercept (if included) is always included in the model.}

\item{scale}{used in the definition of the \code{AIC} statistic for selecting the models,
currently only for \link{lm}, \link{aov} and \link{glm} models. The default value, \code{0}, indicates the
scale should be estimated: see \link{extractAIC}.}

\item{trace}{if positive, information is printed during the running of \code{fastbackward}.
Larger values may give more detailed information. If \code{trace} is greater than 1, then
information about which variables at each step are not considered for removal due
to the bounding algorithm are printed.}

\item{keep}{a filter function whose input is a fitted model object and the associated \code{AIC} statistic,
and whose output is arbitrary. Typically \code{keep} will select a subset of the components
of the object and return them. The default is not to keep anything.}

\item{steps}{the maximum number of steps to be considered. The default is 1000
(essentially as many as required). It is typically used to stop the process early.}

\item{k}{the multiple of the number of degrees of freedom used for the penalty.
Only \code{k = 2} gives the genuine AIC: \code{k = log(n)} is sometimes referred to as BIC or SBC.}

\item{...}{any additional arguments to \link{extractAIC}.}
}
\value{
The stepwise-selected model is returned, with up to two additional components.
There is an "\code{anova}" component corresponding to the steps taken in the search,
as well as a "\code{keep}" component if the \verb{keep=} argument was supplied in the call.
The "\verb{Resid. Dev}" column of the analysis of deviance table refers to a constant
minus twice the maximized log likelihood: it will be a deviance only in cases
where a saturated model is well-defined (thus excluding \code{lm}, \code{aov} and \code{survreg} fits,
for example)
}
\description{
Performs backward elimination by AIC, backward elimination is
performed with a bounding algorithm to make it faster.
}
\details{
The bounding algorithm allows us to avoid fitting models that cannot possibly
provide an improvement in AIC. At a high-level, the algorithm basically
works by identifying important predictors that if they are removed from the
current model then they cannot possibly improve upon the current AIC.

Test statistics, p-values, and confidence intervals from the final selected
model are not reliable due to the selection process. Thus, it is not recommended
to use these quantities.

See more details at \link{step}.
}
\examples{
# Loading in fastbackward package
library(fastbackward)

# example with lm
summary(lm1 <- lm(Fertility ~ ., data = swiss))

## step
slm1 <- step(lm1, direction = "backward")
summary(slm1)
slm1$anova

## fastbackward
slm1 <- fastbackward(lm1)
summary(slm1)
slm1$anova

## fastbackward with trace > 1
slm1 <- fastbackward(lm1, trace = 2)

}
\seealso{
\link{step}, \link{drop1}, and \link{extractAIC}
}
